Improve Quick sort by using circular doubly linked list
In the original implementation of non-recursive Quick sort, we
need to push the head of list and tail of list to the stack, which
will take linear time to get the tail of list by using singly
linked list.

To reduce the time spending finding the tail, I use doubly linked
list instead, which takes constant time to get the tail. I implement
it by using the Linux Kernel API.